requirements.txt
grpcio-tools ~= 1.30
kafka-python ~=2.0.2